feat: initial deprecation of mcp server in favor of only client. still not fully working + needs docs updates
