+++ mirascope/mcp/server.py
@@ -1,6 +1,12 @@
-"""MCP server implementation."""
+"""MCP server implementation.
+
+DEPRECATED: The MCP server implementation is deprecated and will be removed in a future version.
+Mirascope will only implement the client-side of MCP in the future, allowing it to connect to any
+MCP server implementation, even if not built with Mirascope.
+"""
 
 import inspect
+import warnings
 from collections.abc import Awaitable, Callable, Iterable
 from typing import Literal, ParamSpec, cast, overload
 
@@ -34,6 +40,15 @@ from mirascope.core.base._utils._messages_decorator import (
 )
 from mirascope.core.base.prompt import PromptDecorator, prompt_template
 from mirascope.mcp.tools import MCPTool, ToolUseBlock
+
+# Show deprecation warning
+warnings.warn(
+    "The MCPServer implementation is deprecated and will be removed in a future version. "
+    "Mirascope will only implement the client-side of MCP in the future. "
+    "We recommend using the official MCP SDK (e.g. `FastMCP`) for server-side implementations.",
+    DeprecationWarning,
+    stacklevel=2,
+)
 
 _P = ParamSpec("_P")
 
@@ -117,7 +132,12 @@ def _generate_prompt_from_function(fn: Callable) -> Prompt:
 
 
 class MCPServer:
-    """MCP server implementation."""
+    """MCP server implementation.
+
+    DEPRECATED: The MCPServer implementation is deprecated and will be removed in a future version.
+    Mirascope will only implement the client-side of MCP in the future, allowing it to connect to any
+    MCP server implementation, even if not built with Mirascope.
+    """
 
     def __init__(
         self,
@@ -134,6 +154,13 @@ class MCPServer:
         ]
         | None = None,
     ) -> None:
+        warnings.warn(
+            "MCPServer is deprecated and will be removed in a future version. "
+            "Mirascope will only implement the client-side of MCP in the future. "
+            "We recommend using the official MCP SDK (e.g. `FastMCP`) for server-side implementations.",
+            DeprecationWarning,
+            stacklevel=2,
+        )
         self.name: str = name
         self.version: str = version
         self.server: Server = Server(name)
