+++ pyproject.toml
@@ -84,7 +84,7 @@ realtime = [
     "sounddevice>=0.5.1,<1",
     "websockets>=13.1,<14",
 ]
-mcp = ["mcp>=1.0.0"]
+mcp = ["mcp>=1.3.0"]
 xai = ["openai>=1.6.0,<2"]
 
 [tool.uv]
